refactor TC_13.002 | User > Create User > Manage Jenkins > Users
